+++ automatic/io-ninja/tools/chocolateyuninstall.ps1
@@ -2,7 +2,7 @@
 
 $packageArgs = @{
   packageName   = $env:ChocolateyPackageName
-  softwareName  = 'io-ninja*'
+  softwareName  = 'IO Ninja*'
   fileType      = 'msi'
   silentArgs    = "/qn /norestart /l*v `"$($env:TEMP)\$($env:chocolateyPackageName).$($env:chocolateyPackageVersion).MsiInstall.log`""
   validExitCodes= @(0, 3010, 1641)
@@ -13,7 +13,7 @@ $uninstalled = $false
 [array]$key = Get-UninstallRegistryKey @packageArgs
 
 if ($key.Count -eq 1) {
-  $key | % {
+  $key | ForEach-Object {
     $packageArgs['silentArgs'] = "$($_.PSChildName) $($packageArgs['silentArgs'])"
     $packageArgs['file'] = ''
 
@@ -25,5 +25,5 @@ if ($key.Count -eq 1) {
   Write-Warning "$($key.Count) matches found!"
   Write-Warning "To prevent accidental data loss, no programs will be uninstalled."
   Write-Warning "Please alert the package maintainer that the following keys were matched:"
-  $key | % { Write-Warning "- $($_.DisplayName)" }
+  $key | ForEach-Object { Write-Warning "- $($_.DisplayName)" }
 }