+++ automatic/io-ninja/tools/helpers.ps1
@@ -1,9 +1,9 @@
-if (!(Test-Path Function:\Import-Certificate)) {
+﻿if (!(Test-Path Function:\Import-Certificate)) {
   function Import-Certificate([string]$FilePath, [string]$CertStoreLocation) {
     $certSplits = $CertStoreLocation -replace '^cert\:\\' -split '\\'
     if (!($certSplits.Length -eq 2)) { throw "Unexpected certificate storage location" }
 
-    $certutil = Get-Command "certutil" | select -ExpandProperty Path
+    $certutil = Get-Command "certutil" | Select-Object -ExpandProperty Path
 
     if (!$certutil) { throw "Path to certutil was not found" }
 
@@ -13,6 +13,6 @@ if (!(Test-Path Function:\Import-Certificate)) {
       "$(Resolve-Path $FilePath)"
     )
 
-    Start-ChocolateyProcess -ExeToRun $certutil -Statements $arguments
+    Start-ChocolateyProcessAsAdmin -ExeToRun $certutil -Statements $arguments
   }
 }