+++ automatic/io-ninja/Changelog.md
@@ -1,4 +1,4 @@
-# Package changelog for [io-ninja](https://chocolatey.org/packages/io-ninja)
+# ![IO Ninja Changelog](https://img.shields.io/badge/IO%20Ninja-Package%20Changelog-blue.svg?style=for-the-badge)
 
 ## Version: 3.7.4 (2017-06-21)
 - **ENHANCEMENT:** Added changelog for package